- Removed a VariantSourceEntry constructor - Removed a VariantSourceEntryWithSampleNames constructor - Moved mongodb entities to mongodb submodule
