+++ core/pom.xml
@@ -16,4 +16,6 @@
     </properties>
 
 
+
+
 </project>
