+++ mongodb/src/main/java/uk/ac/ebi/eva/commons/mongodb/entities/subdocuments/AnnotationIndexMongo.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.commons.core.models.mongodb.subdocuments;
+package uk.ac.ebi.eva.commons.mongodb.entities.subdocuments;
 
 import org.springframework.data.mongodb.core.mapping.Field;
 import org.springframework.util.Assert;
 import uk.ac.ebi.eva.commons.core.models.IXref;
-import uk.ac.ebi.eva.commons.core.models.mongodb.AnnotationMongo;
+import uk.ac.ebi.eva.commons.mongodb.entities.AnnotationMongo;
 
 import java.util.ArrayList;
 import java.util.Collection;
