+++ mongodb/src/main/java/uk/ac/ebi/eva/commons/mongodb/entities/subdocuments/HgvsMongo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.commons.core.models.mongodb.subdocuments;
+package uk.ac.ebi.eva.commons.mongodb.entities.subdocuments;
 
 import org.springframework.data.mongodb.core.mapping.Field;
