+++ mongodb/src/main/java/uk/ac/ebi/eva/commons/mongodb/repositories/AnnotationMetadataRepository.java
@@ -17,7 +17,7 @@ package uk.ac.ebi.eva.commons.mongodb.repositories;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import uk.ac.ebi.eva.commons.core.models.mongodb.AnnotationMetadataMongo;
+import uk.ac.ebi.eva.commons.mongodb.entities.AnnotationMetadataMongo;
 
 import java.util.List;
