+++ mongodb/src/main/java/uk/ac/ebi/eva/commons/mongodb/repositories/FeatureRepository.java
@@ -16,7 +16,7 @@
 package uk.ac.ebi.eva.commons.mongodb.repositories;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import uk.ac.ebi.eva.commons.core.models.mongodb.FeatureCoordinatesMongo;
+import uk.ac.ebi.eva.commons.mongodb.entities.FeatureCoordinatesMongo;
 
 import java.util.List;
