+++ mongodb/src/main/java/uk/ac/ebi/eva/commons/mongodb/repositories/VariantRepository.java
@@ -21,7 +21,7 @@ package uk.ac.ebi.eva.commons.mongodb.repositories;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.mongodb.repository.Query;
 import uk.ac.ebi.eva.commons.core.models.VariantType;
-import uk.ac.ebi.eva.commons.core.models.mongodb.VariantMongo;
+import uk.ac.ebi.eva.commons.mongodb.entities.VariantMongo;
 
 import java.util.List;
