+++ mongodb/src/main/java/uk/ac/ebi/eva/commons/mongodb/repositories/VariantSourceRepository.java
@@ -19,7 +19,7 @@ import com.google.common.collect.HashBasedTable;
 import com.google.common.collect.Table;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
-import uk.ac.ebi.eva.commons.core.models.mongodb.VariantSourceMongo;
+import uk.ac.ebi.eva.commons.mongodb.entities.VariantSourceMongo;
 
 import java.util.Arrays;
 import java.util.List;
